// lib/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      image_jobs: {
        Row: {
          created_at: string
          error_message: string | null
          feature_id: string
          final_output_url: string | null
          final_seed_used: number | null
          id: string
          is_public: boolean
          provider: string
          provider_job_id: string | null
          provider_model: string | null
          request_params: Json
          status: Database["public"]["Enums"]["job_status"]
          temp_output_url: string | null
          updated_at: string
          user_id: string
        }
        Insert: {
          created_at?: string
          error_message?: string | null
          feature_id: string
          final_output_url?: string | null
          final_seed_used?: number | null
          id?: string
          is_public?: boolean
          provider?: string
          provider_job_id?: string | null
          provider_model?: string | null
          request_params: Json
          status?: Database["public"]["Enums"]["job_status"]
          temp_output_url?: string | null
          updated_at?: string
          user_id: string
        }
        Update: {
          created_at?: string
          error_message?: string | null
          feature_id?: string
          final_output_url?: string | null
          final_seed_used?: number | null
          id?: string
          is_public?: boolean
          provider?: string
          provider_job_id?: string | null
          provider_model?: string | null
          request_params?: Json
          status?: Database["public"]["Enums"]["job_status"]
          temp_output_url?: string | null
          updated_at?: string
          user_id?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never